backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-ai>=1.44.0
python-dotenv>=1.0.0
httpx>=0.27.2
pydantic>=2.10
python-multipart>=0.0.6
